Fix lambda inference CI regressions
