+++ baml_language/crates/baml_ide/Cargo.toml
@@ -22,7 +22,6 @@ baml_compiler_parser = { workspace = true }
 baml_compiler_syntax = { workspace = true }
 baml_db = { workspace = true }
 baml_type = { workspace = true }
-baml_type_runtime = { workspace = true }
 bitflags = { workspace = true }
 la-arena = { workspace = true }
 rowan = { workspace = true }
